Load network data from disk
We write the networks data (and users) to disk every time
generatetestnetworks.js is invoked. The default behaviour is to generate
the networks data.

Generating networks data is complex and time consuming, so for the users
willing to use old data, they can specify -l in the shell script in
order to load networks data from disk.