'Pre' scripts should not be included as part of test Suites
Move 'Pre' directory out of Suites dir.
Default execution of runner.sh is to run all tests in the tests/suites
dir.
Because Pre dir exists in /suites, it was being included in the test
run. Scripts in Pre are designed specifically for data environment
propping, and it does not make sense to view them as test scripts.